agent: set combined channels for eth0 to max allowed
A network interface can use multiple threads handling queues to greatly
increase network latency and throughput on multicore systems with
multiple network streams.

Multiqueue capability is usually set at boot time, but when an interface
is changed at runtime it will stay misconfigured until the software
sets the number of queues appropriately again.

This patch checks the number of queues an interface is capable of handle
and properly sets it via ioctl any time an interface link is updated.

Signed-off-by: Angelo Ruocco <ang@zurich.ibm.com>