vc: adjust default and max CPUs after we move shim
We are setting up the runtime config earlier in configuration
load and are validating the maximum CPU provided in the configuration
file "makes sense" based on how many CPUs are available to the
shim process. At this point in time, however, the shim is bound by
how many CPUs its parent have.

In the event that the upper layer runtime (say containerd) is
running within a constrained cpuset cgroup, this check will return however
many CPUs the upper layer has available. Once the actual pod/sandbox
is started, however, we'll no longer be in this cgroup and may have
a different limitation (or no limitation).

Based on this unexpected behavior, let's make sure that we adjust the
default and maximum number of vCPUs *after* we move into our own cgroup
hierarchy.

Fixes: #9327

Signed-off-by: Eric Ernst <eric_ernst@apple.com>