gha: k8s: DM: Remove state files and fix DM path
Remove any k3s/containerd state files that will cause it to replay devicemapper
transactions.  We do not want this to happen since we are recreating the
underlying storage and want a fresh environment.

Also, fixed the DM path to be consistent and added a dump of the DM devices.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>