gha: fix issue in containerd switching between different snapshotters on CI
we are encountering the issue (https://github.com/kata-containers/kata-containers/issues/8407)
with containerd on CI is likely due to the content digest being missing from the content store,
which can happen when switching between different snapshotters.
To help sort it out on CI, we now clean up related snapshots or images in k8s.io namespace.

Signed-off-by: ChengyuZhu6 <chengyu.zhu@intel.com>