agent: mounts: Mount configfs into the container
configfs is used to get a quote generated, and having this information
available from inside the container (in case the container itself wants
to attest something) is a must.

It's important to note that this is not really needed by Confidential
Containers as that implies the guest-components are being used. However,
that's not the only use-case we should support, and we've seen other
companies already having to maintain a downstream patch (a version of
this very same patch) to have configfs exposed to their attestation
agents.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>