config: Add `memory_overhead` annotation
Add a `memory_overhead` annotation that an external operator running
in an orchestrator such as Kubernetes or OpenShift can set, and
take that value into account when hot-plugging memory.

The reason this mechanism is needed is that an orchestrator may set a
host-side cgroup based on what it knows, and it considers the "pod
overhead" as specified for example by a runtime class instead of the
default memory size as given in the configuration file.

The pod overhead is supposed to estimate how much additional resources
are required by the runtime, and it needs to be kept relatively low in
order to avoid impacting container density too negatively.

For example, consider typical values where:
- the default VM memory in `configuration.toml` is 2048M,
- the pod overhead is 350M,
- you run a container that requests 4096M.

When we create the pod VM, it boots with 2048M. We don't want the pod
overhead in the orchestrator to be 2048M, though, since in most cases,
the VM uses much less than that. The 350M given to the orchestrator
indicates that we expect the VM to consume 350M in addition to what
the workloads need.

We then add the container to the pod. When the `memory_overhead` is
not set, we hot-plug what the container requests, i.e. 4096M. As a
result, the VM is now allocated 6144M, although it typically consumes
less than that. If the container workload exceeds the 4096M allocated
to it, it will be OOM-killed by the guest cgroup, and that's exactly
what we want.

However, if the container workload is close to that limit, say it uses
3900M, and also uses a lot of guest kernel memory resources, e.g. by
doing lots of I/Os, some of that memory (e.g. file buffers) is not
counted towards the workload by the guest cgroups. Thus, we may end up
with a guest that consumes for example 5500M of memory.

This is not compatible with what the orchestrator expects, which is
4096M for the workload + 350M for the overhead. If the orchestrator
sets up a host-side cgroup based on what it knows, it will set the
limit at 4446M, and a VM that tries to consume more than that will be
OOM-killed _by the host_. Since we kill the hypervisor, we end up with
messages that are much harder to understand and may not end up in the
user-accessible orchestrator logs, e.g.

```
host kernel: oom_reaper: reaped process 4152164 (qemu-kvm), now anon-rss:0kB, file-rss:84kB, shmem-rss:33171748kB
```

The proposed solution is to adjust the first memory hotplug(s) to
ensure that the guest memory lines up with the orchestrator's
expectation. In the above example, the first hot-plug would not be
4096M, but 4096M - (2048M - 350M), making the VM size after hot-plug
identical to what the orchestrator expects.

If there is memory pressure, it now happens in the guest, and the OOM
will now happen in the guest and not in the host.

Fixes: #6129

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>