agent: added param to provide auth for guest-pull
fixes #9802

The parameter is required for the `guest-pull` feature. While the
metadata is pulled on the hosting node, the image layers are being
pulled from the registry within the guest by image-rs. The image client
can be initialized with a config.json auth file for OCI registries.

The change adds a param, gated behind `guest-pull` feature, to provide
image-rs with an auth.config file. A simple sanity check for existence
of this provided file is being performed at initialization.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>