workflows/kata-runtime-classes-sync: run only on given files changed
This workflow is meant to check changes in
tools/packaging/kata-deploy/runtimeclasses so let's not run on every and each
PR.

While here, it was gave a name to the job so that on UI it will show up as
'Static checks / kata-deploy-runtime-classes-check' similarly to other
static checks (e.g. 'Static checks / check-kernel-config-version').

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>