static-checks: silencing the kata-check-markdown
The kata-check-markdown command has logged too much messages (at least 3 lines
for each file checked) that on CI makes it difficult to quickly spot a fail
because,When an error occurs and that's often not related with markdown problems,
you need to open the raw logs file to scroll to the end because Github refuses
to displaylong text. Even if one try to enclose the kata-check-markdown in a
folding group, long texts exceeding the maximum allowed will make Github to
(again) refuse to display the full log.

With this change the kata-check-markdown, executed through static-checks.sh, will
have its output silenced by default. But on error, it will print the message then.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>