ci: kbs: Force delete the KBS pod
I've noticed that the KBS pod is getting stuck in "Terminating" state,
which blocks the CI to finish.

This code path is only running on TDX for now, as the coco-qemu-dev runs
do not clean up the KBS bits, so it may be a TDX machine issue or it may
be a KBS issue.

Workarounds: #9885

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>