coco-ci: patch kbs' rego policies in configmap
In recent revisions in kbs' kustomize deployment the policy is backed by
a kustomize-managed ConfigMap. This mount is read-only in the pod, so
attempts to push a new policy into the pod's local fs will not work.

This change adjust the ConfigMap which should be propagated to the kbs
pod.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>